update of libapw

+++ src/libapw/cpp/lapw_load_global.cpp
@@ -67,7 +67,7 @@ extern "C" void FORTRAN(lapw_load_global)(int *lmaxvr_,
     lapw_global.igfft.resize(lapw_global.ngrtot);
     lapw_global.cfunir.resize(lapw_global.ngrtot);
     lapw_global.cfunig.resize(lapw_global.ngrtot);
-    for (unsigned int i = 0; i < lapw_global.ngrtot; i++)
+    for (int i = 0; i < lapw_global.ngrtot; i++)
     {
         lapw_global.cfunig[i] = cfunig_[i];
         lapw_global.cfunir[i] = cfunir_[i];
@@ -88,7 +88,7 @@ extern "C" void FORTRAN(lapw_load_global)(int *lmaxvr_,
     lapw_global.natmcls = *natmcls_;
     lapw_global.ic2ias.resize(lapw_global.natmcls);
     lapw_global.natoms_in_class.resize(lapw_global.natmcls);
-    for (unsigned int ic = 0; ic < lapw_global.ic2ias.size(); ic++)
+    for (int ic = 0; ic < (int)lapw_global.ic2ias.size(); ic++)
     {
         lapw_global.ic2ias[ic] = ic2ias_[ic] - 1;
         lapw_global.natoms_in_class[ic] = natoms_in_class_[ic];
@@ -100,34 +100,28 @@ extern "C" void FORTRAN(lapw_load_global)(int *lmaxvr_,
     lapw_global.L3_gntyry.set_dimensions(lapw_global.lmmaxapw, lapw_global.lmmaxapw);
     lapw_global.L3_gntyry.allocate();
     
-    //lapw_global.L3_gntyry_data.set_dimensions(lapw_global.lmmaxapw, lapw_global.lmmaxapw);
-    //lapw_global.L3_gntyry_data.allocate();
-    
-    for (unsigned int lm1 = 0; lm1 < lapw_global.lmmaxapw; lm1++)
-        for (unsigned int lm2 = 0; lm2 < lapw_global.lmmaxapw; lm2++)
-            for (unsigned int lm3 = 0; lm3 < lapw_global.lmmaxvr; lm3++) 
+    for (int lm1 = 0; lm1 < lapw_global.lmmaxapw; lm1++)
+        for (int lm2 = 0; lm2 < lapw_global.lmmaxapw; lm2++)
+            for (int lm3 = 0; lm3 < lapw_global.lmmaxvr; lm3++) 
                 if (abs(lapw_global.gntyry(lm3, lm1, lm2)) > 1e-14)
                 {
                     lapw_global.L3_gntyry(lm1, lm2).push_back(lm3);
-                    //lapw_global.L3_gntyry_data(lm1, lm2).push_back(lapw_global.gntyry(lm3, lm1, lm2));
                 }
     
     
     lapw_global.omega = *omega_;
     
-    for (unsigned int i = 0; i < lapw_global.species.size(); i++)
+    for (int i = 0; i < (int)lapw_global.species.size(); i++)
         delete lapw_global.species[i];
     lapw_global.species.clear();
 
-    for (unsigned int i = 0; i < lapw_global.atoms.size(); i++)
+    for (int i = 0; i < (int)lapw_global.atoms.size(); i++)
         delete lapw_global.atoms[i];
     lapw_global.atoms.clear();
 
-    for (unsigned int i = 0; i < lapw_runtime.bloch_states.size(); i++)
+    for (int i = 0; i < (int)lapw_runtime.bloch_states.size(); i++)
         delete lapw_runtime.bloch_states[i];
     lapw_runtime.bloch_states.clear();
-    
-    //std::cout << "LDA+U : " << lapw_global.ldapu << std::endl;
 }
